% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepairSwitches.r
\name{basepairSwitches}
\alias{basepairSwitches}
\title{Resolve "AB"-"BA" discrepancies}
\usage{
basepairSwitches(x, filename = NULL)
}
\arguments{
\item{x}{use default \code{x = Populations}}

\item{filename}{what would you like to name the export summary}
}
\description{
\code{readInData()} the \code{readInData()} function checks for any "AB"-"BA" discrepancies during genotype import.
The \code{basepairSwitches()} function can be used to convert any "BA" genotypes to "AB" genotypes
}
\author{
Eric Grau and Mike Ackerman
}

