% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearSession.r
\name{clearSession}
\alias{clearSession}
\title{Clear \pkg{idfgen}}
\usage{
clearSession()
}
\description{
This function clears the R console of all things \pkg{idfgen} related
}
\author{
Eric Grau and Mike Ackerman
}

