% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpBaseline.r
\name{dumpBaseline}
\alias{dumpBaseline}
\title{Export GSI baseline file}
\usage{
dumpBaseline(x, markers, fileType, filename, findMarkers = NULL,
  replaceWith = NULL, removePrefix = NULL, replaceBP = TRUE,
  basePairs = c("A", "C", "G", "T", "-"), replacements = c("1", "2", "3",
  "4", "5"))
}
\arguments{
\item{x}{the list of populations to be included in gsi baseline file}

\item{markers}{the markers to be included in gsi baseline file}

\item{fileType}{the format of the exported gsi baseline file. Can be set to "SPAM", "BAYES", or "gsi_sim"}

\item{filename}{what would you like to name the exported gsi baseline file}

\item{findMarkers}{i don't remember what this does, just leave as the default \code{NULL}}

\item{replaceWith}{something related to \code{findMarkers}. Also leave as the default \code{NULL}}

\item{removePrefix}{this can be used to trim the marker names for "SPAM"}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpBaseline()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpBaseline()} exports a gsi baseline file in BAYES, SPAM, or gsi_sim format
}
\author{
Eric Grau and Mike Ackerman
}

