% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpColony.r
\name{dumpColony}
\alias{dumpColony}
\title{Export COLONY file}
\usage{
dumpColony(x, markers = NULL, filename, errorFile = NULL,
  errorDefaultValues = c(0, 0), replaceBP = TRUE, basePairs = c("A", "C",
  "G", "T", "-"), replacements = c("1", "2", "3", "4", "5"))
}
\arguments{
\item{x}{the list of populations to be included in Colony file}

\item{markers}{the markers to be included in Colony file}

\item{filename}{what would you like to name the exported COLONY file?}

\item{errorFile}{what would you like to name the marker error rate file for COLONY?}

\item{errorDefaultValues}{what would you like your defaule genotype error rate to be for COLONY?}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpColony()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpColony()} exports a file in Colony format
}
\author{
Eric Grau and Mike Ackerman
}

