% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpGenepop.r
\name{dumpGenepop}
\alias{dumpGenepop}
\title{Export Genepop file}
\usage{
dumpGenepop(x, markers = NULL, title, filename, popNames = FALSE,
  replaceBP = TRUE, basePairs = c("A", "C", "G", "T", "-"),
  replacements = c("1", "2", "3", "4", "5"))
}
\arguments{
\item{x}{the list of populations to be included in Genepop file}

\item{markers}{the markers to be included in Genepop file}

\item{title}{what would you like the title of your Genepop file to be}

\item{filename}{what would you like to name the exported Genepop file?}

\item{popNames}{would you like to use population names \code{popNames = TRUE} or individual names \code{popNames = FALSE} as individual
identifiers in the Genepop file}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpGenepop()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpGenepop()} exports a file in Genepop format
}
\author{
Eric Grau and Mike Ackerman
}

