% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpMixture.r
\name{dumpMixture}
\alias{dumpMixture}
\title{Export GSI mixture file}
\usage{
dumpMixture(x, markers, fileType, filename, individuals = NULL,
  findMarkers = NULL, replaceWith = NULL, removePrefix = NULL,
  replaceBP = TRUE, basePairs = c("A", "C", "G", "T", "-"),
  replacements = c("1", "2", "3", "4", "5"))
}
\arguments{
\item{x}{the 'population' to be included in the gsi mixture file}

\item{markers}{the markers to be included in gsi mixture file. NOTE: should be same markers set that was used for the \code{dumpBaselin()}
function}

\item{fileType}{the format of the exported gsi mixture file. Can be set to "SPAM", "BAYES", or "gsi_sim"}

\item{filename}{what would you like to name the exported gsi mixture file}

\item{individuals}{keep as default \code{NULL}}

\item{findMarkers}{i don't remember what this does, just leave as the default \code{NULL}}

\item{replaceWith}{something related to \code{findMarkers}. Also leave as the default \code{NULL}}

\item{removePrefix}{this can be used to trim the marker names for "SPAM"}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpMixture()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpMixture()} exports a gsi mixture file in BAYES, SPAM, or gsi_sim format
}
\author{
Eric Grau and Mike Ackerman
}

