% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpMSToolKit.r
\name{dumpMSToolKit}
\alias{dumpMSToolKit}
\title{Export MSToolKit file}
\usage{
dumpMSToolKit(x, markers = NULL, filename, replaceBP = TRUE,
  basePairs = c("A", "C", "G", "T", "-"), replacements = c("1", "2", "3",
  "4", "5"))
}
\arguments{
\item{x}{the list of populations to be included in MSToolKit file}

\item{markers}{the markers to be included in MSToolKit file}

\item{filename}{what would you like to name the exported MSToolKit file?}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpMSToolKit()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpMSToolKit()} exports a file in MSToolKit format
}
\author{
Eric Grau and Mike Ackerman
}

