% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpPhylip.r
\name{dumpPhylip}
\alias{dumpPhylip}
\title{Export Phylip file}
\usage{
dumpPhylip(x, markers = NULL, filename)
}
\arguments{
\item{x}{the list of populations to be included in Phylip file}

\item{markers}{the markers to be included in Phylip file}

\item{filename}{what would you like to name the exported Phylip file?}
}
\description{
\code{dumpPhylip()} exports a file in Phylip format
}
\author{
Eric Grau and Mike Ackerman
}

