% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSNPPIT.r
\name{dumpSNPPIT}
\alias{dumpSNPPIT}
\title{Export snppit file}
\usage{
dumpSNPPIT(baselinePops, mixturePops, markers, filename, errorFile = NULL,
  errorDefaultValue = NULL, POPCOLUMN_SEX = NULL,
  POPCOLUMN_REPRO_YEARS = NULL, POPCOLUMN_SPAWN_GROUP = NULL,
  OFFSPRINGCOLUMN_BORN_YEAR = NULL, OFFSRPINGCOLUMN_SAMPLE_YEAR = NULL,
  OFFSPRINGCOLUMN_AGE_AT_SAMPLING = NULL)
}
\arguments{
\item{baselinePops}{the list of populations to be used as the parental baseline}

\item{mixturePops}{the populations to be included as potential offspring}

\item{markers}{the list of markers to be used in the snppit file}

\item{filename}{what would you like to name the exported snppit file?}

\item{errorFile}{if you have a .txt file containing the genotype error rates for each of the markers to be used in the snppit file, what is the
name of the file? The .txt should contain 2 columns: 1) the name of each markers and 2) the error rate to be used}

\item{errorDefaultValue}{if you don't have error rates for each of your markers, what would you like snppit to use as the default error
rate}

\item{POPCOLUMN_SEX}{If you would like to include sex information for the parents, which column in your input file contains the sex data?}

\item{POPCOLUMN_REPRO_YEARS}{If you would like to include the year that the parents reproduced, which column in your input file contains the
reporductive years?}

\item{POPCOLUMN_SPAWN_GROUP}{If you would like to include spawn groups for the parents, which column in your input file contains the spawn groups?}

\item{OFFSPRINGCOLUMN_BORN_YEAR}{If you would like to include the year born for offspring, which column contains that information?}

\item{OFFSRPINGCOLUMN_SAMPLE_YEAR}{If you would like to include sampling year for offspring, which column in your input file contains that
information?}

\item{OFFSPRINGCOLUMN_AGE_AT_SAMPLING}{If you would like to include the age at sampling for offspring, which column in your input file
contains that information?}
}
\description{
\code{dumpSNPPIT()} exports a file in SNPPIT format
}
\author{
Eric Grau and Mike Ackerman
}

