% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpStructure.r
\name{dumpStructure}
\alias{dumpStructure}
\title{Export Structure file}
\usage{
dumpStructure(x, markers, filename, useIndNames = TRUE, popNumber = TRUE,
  replaceBP = TRUE, basePairs = c("A", "C", "G", "T", "-", "0"),
  replacements = c("1", "2", "3", "4", "5", "-9"))
}
\arguments{
\item{x}{the list of populations to be included in Genepop file}

\item{markers}{the markers to be included in Genepop file}

\item{filename}{what would you like to name the exported Genepop file?}

\item{useIndNames}{would you like to use individual names \code{useIndNames = TRUE} or population name \code{useIndNames = FALSE} as the
individual identifiers}

\item{popNumber}{would you like to include a column in your Structure file that contains the putative population of origin?}

\item{replaceBP}{would you like to replace the allele names with something else. By default, \code{dumpGenepop()} replaces
c("A","C","G","T","-") with c("1","2","3","4","5"). This can be set to "FALSE" to keep data in the original imported format. Alternatively,
the \code{basePairs} and \code{replacements} arguments can be modified to accommodate any desired import or export format.}

\item{basePairs}{If \code{replaceBP = TRUE}, the nomenclature used for the imported genotype data}

\item{replacements}{If \code{replaceBP = TRUE}, what would you like to replace the alleles with?}
}
\description{
\code{dumpStructure()} exports a file in Structure format
}
\author{
Eric Grau and Mike Ackerman
}

