% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpTable.r
\name{dumpTable}
\alias{dumpTable}
\title{Write a simple table}
\usage{
dumpTable(x, filename, row.names = FALSE, sep = "\\t")
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce \code{x} to a data frame.}

\item{filename}{either a character string naming a file or a \link{connection} open for writing. "" indicates output to the console.}

\item{row.names}{either a logical value indicating whether the row names of \code{x} are to be written along with \code{x}, or a character
vector of row names to be written}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}
}
\description{
\code{dumpTable()} just exports any simple table, perhaps containing the biological data for your individuals
}
\author{
Eric Grau and Mike Ackerman
}
\seealso{
\code{\link[utils]{write.table}}
}

