% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProgeny.r
\name{importProgeny}
\alias{importProgeny}
\title{Import Data from Progeny}
\usage{
importProgeny(inputFile, genotypeStartColumn, popColumn = 1,
  sortBPs = FALSE)
}
\arguments{
\item{inputFile}{the .txt file containing the biological and genotype data to be imported into \pkg{idfgen}. The data exported from
Progeny should be in the following format. Export Format: Wide (One Row per Individual); Identification: Unique ID = Individual Name & Export
Using Pedigree and ID; Allele Format: Two Columns Per Call & Export No Calls as "0"; SNP Coding: A,C,G,T; Allele Call: 142,144,146;
Pedigree Options: Use pedigree names. Any additional fields may be added.}

\item{genotypeStartColumn}{the column # in \code{inputFile} where the genotypes begin. This will vary depending on the number of additional fields added to the
Progeny export.}

\item{popColumn}{which column in \code{inputFile} would you like to use to 'parse'. The default is 1, which is where Progeny exports Pedigree
names}

\item{sortBPs}{the \code{readInData()} function will check for "AB"-"BA" discrepancies. If \code{sortBPs = TRUE}, then any "BA" genotypes will
automatically be converted to "AB"}
}
\description{
\code{readInData()} imports data exported from a Progeny database.
}
\author{
Eric Grau and Mike Ackerman
}

